moved Maven exec resolution specific tests into one test class
